+++ src/neo4j/_data.py
@@ -110,10 +110,7 @@ class Record(tuple, t.Mapping):
             other = t.cast(t.Mapping, other)
             return dict(self) == dict(other)
         else:
-            return False
-
-    def __ne__(self, other: object) -> bool:
-        return not self.__eq__(other)
+            return NotImplemented
 
     def __hash__(self):
         return reduce(xor_operator, map(hash, self.items()))