+++ src/neo4j/_sync/io/_bolt.py
@@ -35,6 +35,7 @@ from ..._deadline import Deadline
 from ..._exceptions import (
     BoltError,
     BoltHandshakeError,
+    SocketDeadlineExceededError,
 )
 from ..._io import BoltProtocolVersion
 from ..._meta import USER_AGENT
@@ -887,6 +888,15 @@ class Bolt:
     def _set_defunct(self, message, error=None, silent=False):
         direct_driver = getattr(self.pool, "is_direct_pool", False)
         user_cancelled = isinstance(error, asyncio.CancelledError)
+        connection_failed = isinstance(
+            error,
+            (
+                ServiceUnavailable,
+                SessionExpired,
+                OSError,
+                SocketDeadlineExceededError,
+            ),
+        )
 
         if not (user_cancelled or self._closing):
             log_call = log.error
@@ -913,6 +923,12 @@ class Bolt:
         if user_cancelled:
             self.kill()
             raise error  # cancellation error should not be re-written
+        if not connection_failed:
+            # Something else but the connection failed
+            # => we're not sure which state we're in
+            # => ditch the connection and raise the error for user-awareness
+            self.close()
+            raise error
         if not self._closing:
             # If we fail while closing the connection, there is no need to
             # remove the connection from the pool, nor to try to close the
@@ -1037,7 +1053,7 @@ def tx_timeout_as_ms(timeout: float) -> int:
         raise err_type(msg) from None
     if timeout < 0:
         raise ValueError("Timeout must be a positive number or 0.")
-    ms = int(round(1000 * timeout))
+    ms = round(1000 * timeout)
     if ms == 0 and timeout > 0:
         # Special case for 0 < timeout < 0.5 ms.
         # This would be rounded to 0 ms, but the server interprets this as