+++ src/neo4j/_sync/io/_pool.py
@@ -297,8 +297,7 @@ class IOPool(abc.ABC):
             )
         except Exception as exc:
             log.debug(
-                "[#%04X]  _: <POOL> check re_auth failed %r auth=%s "
-                "force=%s",
+                "[#%04X]  _: <POOL> check re_auth failed %r auth=%s force=%s",
                 connection.local_port,
                 exc,
                 log_auth,
@@ -878,8 +877,7 @@ class Neo4jPool(IOPool):
         # No readers
         if num_readers == 0:
             log.debug(
-                "[#0000]  _: <POOL> no read servers returned from "
-                "server %s",
+                "[#0000]  _: <POOL> no read servers returned from server %s",
                 address,
             )
             return None