+++ tests/unit/async_/test_driver.py
@@ -728,19 +728,15 @@ async def test_warn_notification_severity_driver_config(
 ) -> None:
     if inspect.isclass(expected) and issubclass(expected, Exception):
         assert min_sev is not ...  # makes no sense to test
-        with (
-            pytest.raises(expected),
-            pytest.warns(PreviewWarning, match="notification warnings"),
-        ):
+        with pytest.raises(expected):
             AsyncGraphDatabase.driver(uri, warn_notification_severity=min_sev)
         return
     if min_sev is ...:
         driver = AsyncGraphDatabase.driver(uri)
     else:
-        with pytest.warns(PreviewWarning, match="notification warnings"):
-            driver = AsyncGraphDatabase.driver(
-                uri, warn_notification_severity=min_sev
-            )
+        driver = AsyncGraphDatabase.driver(
+            uri, warn_notification_severity=min_sev
+        )
     async with driver:
         if min_sev_session is ...:
             session = driver.session()